requirements.txt
fastapi
uvicorn[standard]
streamlit
transformers
sentence-transformers
faiss-cpu
pytesseract
pdf2image
python-multipart
pydantic
langchain
langchain-community
langchain-text-splitters
torch
accelerate
pydantic-settings
python-dotenv
pillow
typing-extensions
asyncio
aiofiles
numpy
